Inline the merkle_channel tests and drop the tests-only directory
crates/iop-prover/src/merkle_channel/ existed only to hold tests.rs. It
was the one place in the repo pairing a foo.rs module with a sibling
foo/ directory, so browsing iop-prover/src/ showed a folder that looked
like a submodule tree and was not one.

Move the 211 lines into a #[cfg(test)] mod tests block in
merkle_channel.rs and remove the directory. The test bodies are
unchanged apart from indentation; `use super::{...}` still resolves to
the same module.
